Separating roi items from manager
